Implement general web service styling
Import fonts 'Atkinson Hyperlegible' and 'Dancing Script'. Define body section of the web service to use 'Atkinson Hyperlegible' and have background color of #FCF8E8. Add 'handwritten-text' class that formats the text of that class to use bold style of 'Dancing Script' font and add the class to the navigation bar title.