Implement some navigation bar styling
Move NavBar.js into its own folder inside the 'components' folder and add NavBar.css for styling the navigation bar. Change the background color of the navigation bar to match the color palette. Add a search icon and make the search bar grow in the flex box to fill empty space in the navigation bar. Add a box shadow to indicate currently active page and show same shadow when hovering over navigation bar items. Add a component that can either show the text 'Login' or a username and a user icon.