Implement registration error for internal server errors and add redirect to login page, implement basic login validation
Add response status 500 sending to POST requests made to '/register' that cause a database error. Successful registrations now redirect to '/login?regRedirect=true' that show a message stating that an account has been created and that the user can now log in. Add basic validation to require both input fields to have content when submitting the login form. Add some basic logic to logging in functionality to either show an error or redirect to landing page based on a placeholder variable.