Add component descriptions to IndividualRecipe.js, pass the recipe id to IndividualRecipe component as props, tweak how recipe data is stored in state
Add component descriptions in comments above components to IndividualRecipe.js. Pass the recipe id as a prop to IndividualRecipe so that it can be accessed through this.props.id for fetching recipe data in the future. Tweak how data is stored in IndividualRecipe state so that only the data is stored and then passed to the components in render instead of storing components itself in the state.