Add input validation to recipe adding form, add more descriptions for components, improve keys, handle file storing in RecipeForm's state
Add 'required' flag to multiple input fields so that the browser enforces the requirement. Add validation for subgroup name input and ingredient input to require that all input fields related to them have content and that the content doesn't contain duplicates. Add more component descriptions in comments to the file. Improve keys so that all keys should be unique and consistent on the page to prevent problems with React rerendering. Handle file storing to store the main recipe image and step images in RecipeForm's state.