Add manual requirement validation to RecipeFrom on top of the validation done by the browser, add ingredient minimum requirement validation, trim input values when applicable
Add manual requirement validation to when the user tries to submit the form. If any validation fails, the error is shown above the appropriate input component and all errors are shown at the bottom above the submit button. Add manual validation to check that ingredients include at least one ingredient in total and that no subgroups are empty. Trim text inputs to remove trailing white spaces when applicable.