Implement login frontend functionality, make navigation bar add the .active class to buttons based on pathname, make user page buttons change active status on click
Implement login frontend functionality to send a POST request to 'localhost:5000/login' with the inputted username and password.  Define cors origin to be 'localhost:3000' and credentials to be true in server index.js. Based on response either show and error message or save the current username into localStorage and redirect to '/recipes'. Navigation bar now checks if an authentication cookie and localStorage username exist and based on that either shows a login button or a button to user page with the username that is saved to localStorage. Navigation bar buttons should now display as active if pathname matches the button's href. User page buttons for showing certain recipes now update the active status when clicked. Removed extra tabs in UserHeader.js code.