Implement basic frontend interface for submitting bookmarks and star rating
Implement basic frontend interface for submitting bookmarks and a star rating for individual recipes. Clicking on the bookmark icon marks the recipe bookmarked in the RecipeHeader's state and alters the color of the bookmark icon. Add star rating interface that saves the inputted star amount in RecipeHeader's state. When a rating is submitted, the star rating input is hidden and the user is shown the rating they have submitted.